Generalize HTTP/2 run variants to ServeConnection
This will allow for many different protocol-level configurations, rather
than just the two represented by 'run' and 'runHttp2'.  In particular,
HTTP/2-only could be supported through these variants, and
HTTP/1-with-upgrade could be a ServeConnection that starts off like
HTTP/1 but hands over the connection to HTTP/2 when the appropriate
Upgrade header is present.