Wait for pushed stream to be sent before returning
This means we can no longer fall victim to the following race:
- Enqueue an OPush on the priority queue and return from pushResponder.
- Send the entirety of the (very short) body and close the stream.
- Dequeue the OPush and ignore it because the stream is closed.

Now if you call pushResponder before finishing the stream, it will have
been decided whether the push will succeed before control returns from
pushResponder.  Of course, the push can still fail if it would exceed
the max concurrency, if pushes are disabled, if the peer resets the
associated stream, or if you call pushResponder after finishing the
response.