Move exception handling to top of thread (fixes #613)
This is a long-standing bug, but hardly manifested until
3eb586636b7f0a9a8e4b4ad456c16efe23847c9e landed. With that change, the
bracket function calls outside of the unmask were likely to receive an
async timeout exception. Without a wrapping call to settingsOnException,
these would be printed to the console, which isn't desired.

Moving the handle call to the top of the thread ensures that any
exception which occurs inside the thread has the handler applied.